chore: move resolve to another module
